requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
anyio==3.7.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
dataclasses-json==0.5.14
Django==4.2.5
exceptiongroup==1.1.3
fastapi==0.103.1
filelock==3.12.4
frozenlist==1.4.0
fsspec==2023.9.1
greenlet==2.0.1
h11==0.14.0
huggingface-hub==0.17.1
idna==3.4
image==1.5.33
Jinja2==3.1.2
langchain==0.0.292
langsmith==0.0.37
MarkupSafe==2.1.3
marshmallow==3.20.1
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
numexpr==2.8.6
numpy==1.25.2
openai==0.28.0
packaging==23.1
Pillow==10.0.1
pydantic==1.10.12
pyTelegramBotAPI==4.13.0
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
safetensors==0.3.3
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.20
sqlparse==0.4.4
starlette==0.27.0
sympy==1.12
telebot==0.0.5
tenacity==8.2.3
tokenizers==0.13.3
torch==2.0.1
tqdm==4.66.1
transformers==4.33.2
typing-inspect==0.9.0
typing_extensions==4.7.1
urllib3==2.0.4
uvicorn==0.23.2
yarl==1.9.2
